Change faerie dragons to fairies
This commit changes almost all references to faerie dragons in the code
to fairies (the only place I haven't changed is the tiles files). We
felt that the species was most definitely not a dragon, due to its low
hp, small size, lack of a breath and a different scales mutation to
others. Simply being a fairy fits this role much much better.

I haven't yet compiled this, so if something goes wrong, it's probably
this huge commit.