Make shields a viable option for fairies
The -3 aptitude is worth training for a buckler sometimes, but
especially given how shield plus magical staff is not possible, and how
they already start with bonus shield, it would have been completely
unreasonable ever to train 21 skill for a regular shield. This commit
makes the aptitude +2, to make this a consideration in some
circumstances.

This may be reverted if the consensus becomes to make them tiny.