Make the fire dragon scales of Xtahua actually drop from Xtahua
The order and parameters with which items() and make_item_unrandart()
are called are very important, and if they're not done correctly we get
all sorts of obscure asserts failing in a variety of different files.

This commit removes the fire dragon scales of Xtahua generating randomly
and instead makes them drop, if a corpse is created (50% chance), from
Xtahua.